cleaning
solve https://github.com/Open-Building-Management/ladybug_codes/issues/137

and https://github.com/Open-Building-Management/ladybug_codes/issues/138
